shareholders_application/requirements.txt
#pip==24.0
asgiref==.7.2
Django==4.2.10
django-countries==7.5.1
django-crispy-forms==2.1
djangorestframework==3.14.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2024.1
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0